requirements.txt
scikit-image
numpy
pytorch
torchvision